requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
requests==2.31.0
bs4==0.0.1
lxml==4.9.3
aiohttp>=3.9.0